requirements.txt
fastapi
uvicorn
openai
google-generativeai
python-dotenv

